Fixed error messages when uninstalling with SQLite

+++ updates/builder_table_update_dizoo_slider_slides.php
@@ -23,17 +23,68 @@ class BuilderTableUpdateDizooSliderSlides extends Migration
     
     public function down()
     {
-        Schema::table('dizoo_slider_slides', function($table)
-        {
-            $table->dropColumn('button_1_url');
-            $table->dropColumn('button_2_url');
-            $table->string('subtitle', 40)->nullable(false)->change();
-            $table->string('description', 210)->nullable(false)->change();
-            $table->boolean('button_1_active')->nullable(false)->change();
-            $table->string('button_1_text', 15)->nullable(false)->change();
-            $table->string('button_1_color', 7)->nullable(false)->default(null)->change();
-            $table->boolean('button_2_active')->default(null)->change();
-            $table->string('button_2_text', 15)->nullable(false)->change();
-        });
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_2_url')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->dropColumn('button_2_url');
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_1_url')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->dropColumn('button_1_url');
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'subtitle')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->string('subtitle', 40)->nullable(false)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'description')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->string('description', 210)->nullable(false)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_1_active')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->boolean('button_1_active')->nullable(false)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_1_text')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->string('button_1_text', 15)->nullable(false)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_1_color')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->string('button_1_color', 7)->nullable(false)->default(null)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_2_active')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->boolean('button_2_active')->default(null)->change();
+            });
+        }
+
+        if (Schema::hasColumn('dizoo_slider_slides', 'button_2_text')) {
+            Schema::table('dizoo_slider_slides', function($table)
+            {
+                $table->string('button_2_text', 15)->nullable(false)->change();
+            });
+        }
     }
 }
